# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Workflow Agent Service API
  version: 0.1.0
  description: Versioned public HTTP contract for the workflow and Agent platform.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
security:
  - cookieAuth: []
...
paths:
...
  /workflow/node-assignment-rule:
    get:
      tags: [Workflow]
      operationId: pageAssignmentRules
      summary: Page node assignment rules
      parameters:
        - { $ref: '#/components/parameters/PageNum' }
        - { $ref: '#/components/parameters/PageSize' }
        - { name: processDefinitionKey, in: query, schema: { type: string } }
        - { name: processDefinitionId, in: query, schema: { type: string } }
        - { name: version, in: query, schema: { type: integer } }
        - { name: taskDefinitionKey, in: query, schema: { type: string } }
        - { name: variableName, in: query, schema: { type: string } }
        - { name: assignmentType, in: query, schema: { type: string } }
        - { name: emptyUserStrategy, in: query, schema: { type: string } }
      responses:
        '200': { description: Assignment rules, content: { application/json: { schema: { $ref: '#/components/schemas/AssignmentRulePageApiResponse' } } } }
    post:
      tags: [Workflow]
      operationId: createAssignmentRule
      summary: Create a node assignment rule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignmentRuleCommand' }
      responses:
        '200': { description: Assignment rule created, content: { application/json: { schema: { $ref: '#/components/schemas/AssignmentRuleApiResponse' } } } }
  /workflow/node-assignment-rule/inherit:
    post:
      tags: [Workflow]
      operationId: inheritAssignmentRules
      summary: Inherit assignment rules from a process definition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignmentRuleInheritCommand' }
      responses:
        '200': { description: Rules inherited, content: { application/json: { schema: { $ref: '#/components/schemas/AssignmentRuleInheritApiResponse' } } } }
  /workflow/node-assignment-rule/{id}:
    post:
      tags: [Workflow]
      operationId: updateAssignmentRule
      summary: Update a node assignment rule
      parameters: [{ $ref: '#/components/parameters/NumericId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignmentRuleCommand' }
      responses:
        '200': { description: Assignment rule updated, content: { application/json: { schema: { $ref: '#/components/schemas/VoidApiResponse' } } } }
    delete:
      tags: [Workflow]
      operationId: deleteAssignmentRule
      summary: Delete a node assignment rule
      parameters: [{ $ref: '#/components/parameters/NumericId' }]
      responses:
        '200': { description: Assignment rule deleted, content: { application/json: { schema: { $ref: '#/components/schemas/VoidApiResponse' } } } }
...
  /operations/agent-completion-events/dead-letters:
    get:
      tags: [Operations]
      operationId: listAgentCompletionDeadLetters
      summary: List dead-letter Agent completion events
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
      responses:
        '200': { description: Dead-letter events, content: { application/json: { schema: { $ref: '#/components/schemas/DeadLetterListApiResponse' } } } }
  /operations/agent-completion-events/{eventId}/replay:
    post:
      tags: [Operations]
      operationId: replayAgentCompletionEvent
      summary: Replay a dead-letter Agent completion event
      parameters:
        - { name: eventId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Event scheduled for replay, content: { application/json: { schema: { $ref: '#/components/schemas/VoidApiResponse' } } } }
  /operations/agent-completion-events/{eventId}/ignore:
    post:
      tags: [Operations]
      operationId: ignoreAgentCompletionEvent
      summary: Acknowledge a dead-letter Agent completion event without replaying it
      parameters:
        - { name: eventId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IgnoreAgentCompletionEventCommand' }
      responses:
        '200': { description: Event ignored, content: { application/json: { schema: { $ref: '#/components/schemas/VoidApiResponse' } } } }
components:
  parameters:
    PageNum:
      name: pageNum
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    NumericId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: WORKFLOW_ACCESS_TOKEN
  schemas:
    IgnoreAgentCompletionEventCommand:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 1, maxLength: 1000 }
    ApiResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string }
        message: { type: string }
        data: {}
        traceId: { type: string }
    VoidApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: object, nullable: true } }
...
    AssignmentRuleApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AssignmentRule' } }
    AssignmentRulePageApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AssignmentRulePage' } }
    AssignmentRuleInheritApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AssignmentRuleInheritResult' } }
...
    DeadLetterListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: array, items: { $ref: '#/components/schemas/DeadLetterEvent' } } }
...
    PageMetadata:
      type: object
      required: [total, pageNum, pageSize, records]
      properties:
        total: { type: integer, format: int64 }
        pageNum: { type: integer }
        pageSize: { type: integer }
        records: { type: array, items: {} }
...
    AssignmentRule:
      type: object
      required: [id, processDefinitionId, processDefinitionKey, version, taskDefinitionKey, priority, assignmentType, enabled]
      properties:
        id: { type: integer, format: int64 }
        tenantId: { type: string }
        processDefinitionId: { type: string }
        processDefinitionKey: { type: string }
        version: { type: integer }
        taskDefinitionKey: { type: string }
        priority: { type: integer }
        assignmentType: { type: string }
        emptyUserStrategy: { type: string }
        enabled: { type: boolean }
        description: { type: string, nullable: true }
        conditions: { type: array, items: { type: object } }
        targets: { type: array, items: { type: object } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AssignmentRulePage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties: { records: { type: array, items: { $ref: '#/components/schemas/AssignmentRule' } } }
    AssignmentRuleInheritResult:
      type: object
      required: [targetProcessDefinitionId, targetVersion, copiedCount, skippedCount, skippedReasons]
      properties:
        sourceProcessDefinitionId: { type: string, nullable: true }
        sourceVersion: { type: integer, nullable: true }
        targetProcessDefinitionId: { type: string }
        targetVersion: { type: integer }
        copiedCount: { type: integer }
        skippedCount: { type: integer }
        skippedReasons: { type: array, items: { type: string } }
...
    DeadLetterEvent:
      type: object
      required: [eventId, tenantCode, aggregateId, attemptCount, createdAt]
      properties:
        eventId: { type: string, format: uuid }
        tenantCode: { type: string }
        aggregateId: { type: string }
        attemptCount: { type: integer }
        lastError: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        deadLetteredAt: { type: string, format: date-time, nullable: true }
        processInstanceId: { type: string, nullable: true }
        executionId: { type: string, nullable: true }
...
    AssignmentRuleCommand:
      type: object
      required: [processDefinitionId, taskDefinitionKey, assignmentType]
      properties:
        processDefinitionId: { type: string }
        processDefinitionKey: { type: string }
        version: { type: integer }
        taskDefinitionKey: { type: string }
        assignmentType: { type: string }
        targets: { type: array, items: { type: object, additionalProperties: true } }
        conditions: { type: array, items: { type: object, additionalProperties: true } }
    AssignmentRuleInheritCommand:
      type: object
      required: [processDefinitionId]
      properties:
        processDefinitionId: { type: string }
...